chore: delete Epic-era post-merge, merge-orchestrator, merge-watcher and pre-push residue (refs #4545)
Five clusters the v2.0.0 cutover left as callees with no callers:

- post-merge/phases/ (6 modules) — only importer was one test.
- git-merge-orchestrator.js — kept alive by 4 dedicated test files; v2
  lands exclusively through squash-merge PRs.
- MergeWatcher — the one genuine partial. The listener class is dead, but
  the live close path imported DEFAULT_INTERVAL_SECONDS,
  DEFAULT_MAX_BUDGET_SECONDS and deriveChecksStatus from it. Relocated
  those to lib/orchestration/merge-poll.js, a home the close path owns,
  before deleting the module rather than after a red run.
- Emitter-less lifecycle schemas (Epic-era + acceptance-reconciler sets).
  epic.watch.* stays: it carries an epic.* name but watcher.js emits it.
  Archived ledgers stay readable — only bus.emit() validates, on write.
- check-prepush-recovery.js and its hook spawn. The escape hatch required
  an epic/<id> ref to open, so it could never fire under the Story-only
  model, yet the hook spawned a node process per push to evaluate it.
  The coverage + CRAP gate is now unconditional.
