--- .agents/schemas/lifecycle/acceptance.reconcile.ok.schema.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/acceptance.reconcile.ok.schema.json",
-  "title": "acceptance.reconcile.ok",
-  "description": "Emitted by AcceptanceReconciler when reconciliation passes. Finalizer subscribes only to this event to gate the push and PR create.",
-  "type": "object",
-  "required": ["baseRead"],
-  "properties": {
-    "baseRead": { "type": "boolean" },
-    "reason": { "type": "string" }
-  },
-  "additionalProperties": false
-}
