--- .agents/schemas/lifecycle/acceptance.reconcile.skipped.schema.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/acceptance.reconcile.skipped.schema.json",
-  "title": "acceptance.reconcile.skipped",
-  "description": "Emitted by AcceptanceReconciler when reconciliation is skipped because the linked Acceptance Spec declares zero AC IDs ('empty-spec'). Story #2893 split the `acceptance::n-a` waiver path out to `acceptance.reconcile.waived` so the Finalizer can route waived Epics through to PR creation while empty-spec Epics still terminate without a PR. The reason field is required so the listener never silently no-ops.",
-  "type": "object",
-  "required": ["baseRead", "reason"],
-  "properties": {
-    "baseRead": { "type": "boolean" },
-    "reason": { "type": "string", "minLength": 1 }
-  },
-  "additionalProperties": false
-}
