--- .agents/schemas/lifecycle/epic.cleanup.end.schema.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.cleanup.end.schema.json",
-  "title": "epic.cleanup.end",
-  "description": "Emitted by Cleaner when branch cleanup and temp archive complete.",
-  "type": "object",
-  "required": ["epicId"],
-  "properties": {
-    "epicId": { "type": "integer", "minimum": 1 }
-  },
-  "additionalProperties": false
-}
