--- .agents/schemas/lifecycle/epic.complete.schema.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.complete.schema.json",
-  "title": "epic.complete",
-  "description": "Terminal event for a successful Epic run, emitted by Cleaner once the temp archive + branch reap complete. NotifyDispatcher subscribes to this event to fan the completion out to the curated webhook channel; the agent::done label flip is driven by the post-merge close path.",
-  "type": "object",
-  "required": ["epicId", "prUrl"],
-  "properties": {
-    "epicId": { "type": "integer", "minimum": 1 },
-    "prUrl": { "type": "string", "format": "uri", "minLength": 1 }
-  },
-  "additionalProperties": false
-}
