--- .agents/schemas/lifecycle/epic.finalize.end.schema.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.finalize.end.schema.json",
-  "title": "epic.finalize.end",
-  "description": "Emitted by Finalizer after the PR is opened (or short-circuited by gh pr list --head idempotency check).",
-  "type": "object",
-  "required": ["epicId", "prUrl"],
-  "properties": {
-    "epicId": { "type": "integer", "minimum": 1 },
-    "prUrl": { "type": "string", "format": "uri", "minLength": 1 }
-  },
-  "additionalProperties": false
-}
