--- .agents/schemas/lifecycle/epic.finalize.start.schema.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.finalize.start.schema.json",
-  "title": "epic.finalize.start",
-  "description": "Emitted by Finalizer at the start of the finalize phase (fast-forward / hotspot / baseline / push / gh pr create).",
-  "type": "object",
-  "required": ["epicId"],
-  "properties": {
-    "epicId": { "type": "integer", "minimum": 1 }
-  },
-  "additionalProperties": false
-}
