--- .agents/schemas/lifecycle/epic.merge.armed.schema.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.merge.armed.schema.json",
-  "title": "epic.merge.armed",
-  "description": "Emitted by AutomergeArmer after gh pr merge --auto --squash --delete-branch succeeds. Must be preceded by epic.merge.ready from the same run (merge-gate ordering invariant).",
-  "type": "object",
-  "required": ["prUrl"],
-  "properties": {
-    "prUrl": { "type": "string", "format": "uri", "minLength": 1 },
-    "epicId": { "type": "integer", "minimum": 1 }
-  },
-  "additionalProperties": false
-}
