--- .agents/schemas/lifecycle/epic.merge.blocked.schema.json
@@ -1,14 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.merge.blocked.schema.json",
-  "title": "epic.merge.blocked",
-  "description": "Emitted by AutomergePredicate when the Epic is NOT safe to auto-merge. AutomergeArmer never sees this event; PR stays disarmed.",
-  "type": "object",
-  "required": ["prUrl", "reason"],
-  "properties": {
-    "prUrl": { "type": "string", "format": "uri", "minLength": 1 },
-    "reason": { "type": "string", "minLength": 1 },
-    "epicId": { "type": "integer", "minimum": 1 }
-  },
-  "additionalProperties": false
-}
