--- .agents/schemas/lifecycle/epic.merge.ready.schema.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.merge.ready.schema.json",
-  "title": "epic.merge.ready",
-  "description": "Emitted SOLELY by AutomergePredicate (on epic.automerge.start in production, or epic.watch.end on the test-only Watcher path) once CI is confirmed green and the disqualification gate passes. The ONLY event AutomergeArmer subscribes to. Story #3367: the Finalizer no longer emits this event — emitting it from finalize cascaded epic.close.end synchronously into the auto-merge arm + branch reap, bypassing the predicate gate and deleting the epic branch before the PR merged.",
-  "type": "object",
-  "required": ["prUrl"],
-  "properties": {
-    "prUrl": { "type": "string", "format": "uri", "minLength": 1 },
-    "reason": { "type": "string" },
-    "epicId": { "type": "integer", "minimum": 1 },
-    "prNumber": { "type": "integer", "minimum": 1 }
-  },
-  "additionalProperties": false
-}
