--- .agents/schemas/lifecycle/epic.plan.start.schema.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.plan.start.schema.json",
-  "title": "epic.plan.start",
-  "description": "Emitted when the runner begins planning waves for an Epic.",
-  "type": "object",
-  "required": ["epicId"],
-  "properties": {
-    "epicId": { "type": "integer", "minimum": 1 }
-  },
-  "additionalProperties": false
-}
