--- .agents/schemas/lifecycle/epic.snapshot.end.schema.json
@@ -1,16 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/dsj1984/mandrel/blob/main/.agents/schemas/lifecycle/epic.snapshot.end.schema.json",
-  "title": "epic.snapshot.end",
-  "description": "Emitted when the snapshot phase finishes; carries enumerated story IDs the Epic owns.",
-  "type": "object",
-  "required": ["epicId", "storyIds"],
-  "properties": {
-    "epicId": { "type": "integer", "minimum": 1 },
-    "storyIds": {
-      "type": "array",
-      "items": { "type": "integer", "minimum": 1 }
-    }
-  },
-  "additionalProperties": false
-}
