+++ tests/story-off-branch-e2e.test.js
@@ -1,15 +1,17 @@
 /**
  * Off-branch end-to-end regression baseline (Story #676 / Task #687).
  *
- * Exercises the worktreeIsolation=false codepath through three checkpoints
- * the task description calls out as "must-not-regress":
+ * Exercises the worktreeIsolation=false codepath through the checkpoints the
+ * task description calls out as "must-not-regress":
  *
  *   1. Story init startup log emits the env-aware [ENV] lines and routes
  *      through bootstrapBranch (not bootstrapWorktree).
  *   2. WorktreeManager lifecycle methods short-circuit with no fs/git calls
  *      when constructed with `enabled: false`.
- *   3. The post-merge `worktreeReapPhase` returns `skipped-disabled` and
- *      does not emit orphan-worktree friction.
+ *
+ * A third checkpoint over the Epic-era post-merge `worktreeReapPhase` was
+ * dropped in Story #4545 along with that phase directory — it had no
+ * production importer, so the test was the module's only caller.
  *
  * The captured log lines form the regression baseline — any future change
  * that lets an undefined-path warning or orphan-worktree message leak onto
@@ -21,7 +23,6 @@
 import assert from 'node:assert/strict';
 import test from 'node:test';
 import { resolveRuntime } from '../.agents/scripts/lib/config-resolver.js';
-import { worktreeReapPhase } from '../.agents/scripts/lib/orchestration/post-merge/phases/worktree-reap.js';
 import { WorktreeManager } from '../.agents/scripts/lib/worktree-manager.js';
 
 const SILENT_LOGGER = { info() {}, warn() {}, error() {} };
@@ -33,29 +34,6 @@ function makeFailingGit() {
     );
   };
   return { gitSync: fail, gitSpawn: fail };
-}
-
-function captureProgress() {
-  const lines = [];
-  const fn = (level, msg) => lines.push(`[${level}] ${msg}`);
-  return { fn, lines };
-}
-
-function assertNoForbiddenTokens(lines) {
-  const forbidden = [
-    /\bundefined\b/i,
-    /orphan-worktree/i,
-    /still-registered/i,
-    /stillRegistered/i,
-  ];
-  for (const line of lines) {
-    for (const pat of forbidden) {
-      assert.ok(
-        !pat.test(line),
-        `Off-branch log line must not match ${pat}: "${line}"`,
-      );
-    }
-  }
 }
 
 test('off-branch e2e: AP_WORKTREE_ENABLED=false routes through resolveRuntime as env-override', () => {
@@ -110,33 +88,6 @@ test('off-branch e2e: WorktreeManager lifecycle methods perform zero git/fs work
     skipped: [],
     skippedReason: 'isolation-disabled',
   });
-});
-
-test('off-branch e2e: worktreeReapPhase returns skipped-disabled without emitting orphan-worktree friction', async () => {
-  const progress = captureProgress();
-
-  // Friction signals land on disk as NDJSON via signals-writer.appendSignal
-  // (Epic #1030 Story #1042). The disabled-isolation path bails out before
-  // emitReapFailureFriction is ever reached, so the absence of any signal
-  // write under `temp/epic-100/story-101/` is the post-cutover invariant.
-  const state = await worktreeReapPhase({
-    delivery: { worktreeIsolation: { enabled: false } },
-    storyId: 101,
-    epicId: 100,
-    epicBranch: 'epic/100',
-    repoRoot: process.cwd(),
-    progress: progress.fn,
-    logger: SILENT_LOGGER,
-  });
-
-  assert.equal(state.status, 'skipped-disabled');
-  assert.equal(state.path, null);
-  assert.equal(state.reason, null);
-  assert.equal(state.method, null);
-  assert.deepEqual(progress.lines, [
-    '[WORKTREE] ⏭️ Skipping worktree reap (worktree isolation disabled)',
-  ]);
-  assertNoForbiddenTokens(progress.lines);
 });
 
 test('off-branch e2e: regression baseline log shape matches expected operator surface', () => {
